refactor(glsl): extract debug cook overlay
