+++ ui/src/lib/components/CanvasPreviewLayout.svelte
@@ -1,8 +1,7 @@
 <script lang="ts">
   import type { Snippet } from 'svelte';
-  import { Activity } from '@lucide/svelte/icons';
+  import CookDebugOverlay from './CookDebugOverlay.svelte';
   import ObjectPreviewLayout from './ObjectPreviewLayout.svelte';
-  import * as Tooltip from './ui/tooltip';
   import type { SettingsSchema } from '$lib/settings';
   import type { ExtraMenuItem } from './object-preview-menu-actions';
   import { previewBackgroundColor } from '../../stores/renderer.store';
@@ -103,30 +102,6 @@
     class?: string;
   } = $props();
 
-  let showCookDebug = $state(cookDebugVisible);
-
-  $effect(() => {
-    showCookDebug = cookDebugVisible;
-  });
-
-  const cookStatusLabel = $derived(cookStatus?.status ?? 'cached');
-  const cookReasons = $derived.by(() => {
-    const visibleReasons = (cookStatus?.lastCookReasons ?? []).filter(
-      (reason) => reason !== 'first-frame'
-    );
-
-    return visibleReasons.join(', ') || 'none';
-  });
-  const cookTime = $derived(
-    cookStatus?.lastCookTimeMs == null ? '--' : `${cookStatus.lastCookTimeMs.toFixed(2)}ms`
-  );
-  const cookStatusClass = $derived.by(() => {
-    if (cookStatusLabel === 'cooked') return 'text-emerald-300';
-    if (cookStatusLabel === 'paused') return 'text-amber-300';
-
-    return 'text-zinc-300';
-  });
-
   // Build the interaction class string based on individual flags
   const interactionClass = $derived.by(() => {
     if (!nodrag && !nopan && !nowheel) return 'cursor-move';
@@ -191,40 +166,7 @@
           : `background-color:${$previewBackgroundColor};${pixelated ? 'image-rendering:pixelated;' : ''}${style}`}
       ></canvas>
 
-      {#if showCookDebugOption}
-        <div class="absolute right-1 bottom-1 flex items-end gap-1">
-          {#if showCookDebug}
-            <div
-              class="rounded border border-zinc-700/80 bg-zinc-950/90 px-2 py-1 font-mono text-[10px] leading-tight text-zinc-400 shadow"
-            >
-              <div class="flex gap-2">
-                <span class={cookStatusClass}>{cookStatusLabel}</span>
-                <span>{cookTime}</span>
-              </div>
-              <div class="max-w-48 truncate">reason: {cookReasons}</div>
-              <div>{cookStatus?.cookedFrames ?? 0} cooks</div>
-            </div>
-          {/if}
-
-          <Tooltip.Root>
-            <Tooltip.Trigger>
-              <button
-                class={[
-                  'cursor-pointer rounded border border-zinc-700/80 bg-zinc-950/80 p-1 shadow hover:bg-zinc-800',
-                  showCookDebug ? 'text-emerald-300' : 'text-zinc-300'
-                ]}
-                aria-label={showCookDebug ? 'Hide cook debug' : 'Show cook debug'}
-                onclick={() => (showCookDebug = !showCookDebug)}
-              >
-                <Activity class="h-3.5 w-3.5" />
-              </button>
-            </Tooltip.Trigger>
-            <Tooltip.Content
-              >{showCookDebug ? 'Hide cook debug' : 'Show cook debug'}</Tooltip.Content
-            >
-          </Tooltip.Root>
-        </div>
-      {/if}
+      <CookDebugOverlay enabled={showCookDebugOption} visible={cookDebugVisible} {cookStatus} />
     </div>
   {/snippet}
 </ObjectPreviewLayout>
